refactoring CookieCheck function
