+++ components/snackbar/styled-snackbar.js
@@ -13,7 +13,7 @@ display: ${props => props.display ? 'block': 'none' };
 
 
 .cookieMess_container {    
-    background: rgba(249,249,249,0.95);
+    background: ${props => props.backGround ? props.backGround : "none"};
     border: 1px solid #ccc;
     border-radius: 5px;
     box-shadow: 0px 7px 15px rgb(85 85 85 / 10%);
@@ -22,7 +22,7 @@ display: ${props => props.display ? 'block': 'none' };
     align-items: center;
     margin: 0 auto;
     padding: 16px;
-    width: 544px;
+    width: ${props => props.width ? props.width : "100%"};;
 
         button {
             transition-duration: 0.3s;
